Extract `LinkOptions` component
